Bugfixes for squeryl-record: Index declaration and implicit conversions for fields of objects that are not retrieved in the current query
Closes #840
Closes #843

Added the traits SquerylRecordNumericalExpression and SquerylRecordNonNumericalExpression that are now mixed into the expressions returned
from the implicit conversions. With this trait, defineAs can be used instead of "is", which is ambiguous for record fields.
Moreover, the case that FieldReferenceLinker.takeLastAccessedFieldReference returns None is now handled correctly in the implicit conversions.
Added some test cases for showing that "defineAs" can be used instead of "is" in the schema declaration and that fields from
earlier retrieved records can be used in other expressions now.
There were also issues with Timestamp fields, which are resolved by using Timestamp instead of Date in the implicit conversion for DateTimeField.